Separate applications for API and UI, same repositiory
