+++ api/index.js
@@ -1,23 +1,22 @@
 const express = require('express')
-const next = require('next')
 const {graphqlExpress, graphiqlExpress} = require('graphql-server-express')
+const cors = require('cors')
 const bodyParser = require('body-parser')
-const pCall = require('../util/pCall')
-const schema = require('../schema')
+const schema = require('./schema/index')
+const pCall = require('./lib/pCall')
 
 main()
 
 async function main() {
-    const port = parseInt(process.env.PORT, 10) || 3000
+    const port = parseInt(process.env.PORT, 10) || 8080
     const dev = process.env.NODE_ENV !== 'production'
 
-    const app = next({dev})
-    await app.prepare()
-
     const server = express()
+    server.use(cors())
     server.use('/graphql', bodyParser.json(), graphqlExpress({schema}))
-    server.use('/graphiql', graphiqlExpress({endpointURL: '/graphql'}))
-    server.get('*', app.getRequestHandler())
+    if (dev) {
+        server.use('/graphiql', graphiqlExpress({endpointURL: '/graphql'}))
+    }
     await pCall(server, 'listen', port)
 
     console.log(`> Ready on http://localhost:${port}`)
