+++ api/nodemon.json
@@ -1,0 +1,9 @@
+{
+  "verbose": true,
+  "watch": [
+    "lib/**/*.*",
+    "schema/**/*.*",
+    "index.js"
+  ],
+  "ext": "*"
+}
