+++ lerna.json
@@ -1,0 +1,8 @@
+{
+  "lerna": "2.2.0",
+  "packages": [
+    "api",
+    "ui"
+  ],
+  "version": "1.0.0"
+}
