--- nodemon.json
@@ -1,11 +1,0 @@
-{
-  "verbose": true,
-  "watch": [
-    "package.json",
-    "package-lock.json",
-    "next.config.js",
-    "server",
-    "schema"
-  ],
-  "ext": "*"
-}
