+++ package.json
@@ -1,30 +1,15 @@
 {
   "name": "next-with-apollo",
   "version": "1.0.0",
-  "author": "Matthew Brunetti <mbrunetti@manitoulintransport.com>",
-  "license": "UNLICENSED",
-  "private": true,
+  "author": "Matthew Francis Brunetti",
+  "license": "MIT",
   "scripts": {
-    "dev": "nodemon server",
-    "build": "next build",
-    "start": "cross-env NODE_ENV=production node server"
-  },
-  "dependencies": {
-    "apollo-local-query": "^0.3.0",
-    "body-parser": "^1.18.1",
-    "cross-env": "^5.0.5",
-    "express": "^4.15.4",
-    "graphql": "^0.9.3",
-    "graphql-server-express": "^1.1.2",
-    "graphql-tools": "^1.2.2",
-    "isomorphic-fetch": "^2.2.1",
-    "next": "^3.2.2",
-    "prop-types": "^15.5.8",
-    "react": "^15.5.4",
-    "react-apollo": "^1.1.3",
-    "react-dom": "^15.5.4"
+    "clean": "lerna clean",
+    "bootstrap": "lerna bootstrap",
+    "test": "lerna run test",
+    "dev": "lerna run dev --parallel"
   },
   "devDependencies": {
-    "nodemon": "^1.12.1"
+    "lerna": "^2.2.0"
   }
 }
