+++ ui/config.js
@@ -1,0 +1,3 @@
+module.exports = {
+    apiBaseUrl: 'http://localhost:8080' // API Server URL (must be absolute)
+}
