+++ ui/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "ui",
+  "version": "1.0.0",
+  "scripts": {
+    "test": "echo \"No tests to run.\" && exit 0",
+    "dev": "next",
+    "build": "next build",
+    "start": "cross-env NODE_ENV=production next start"
+  },
+  "dependencies": {
+    "cross-env": "^5.0.5",
+    "graphql": "^0.9.3",
+    "isomorphic-fetch": "^2.2.1",
+    "next": "^3.2.2",
+    "prop-types": "^15.5.8",
+    "react": "^15.5.4",
+    "react-apollo": "^1.1.3",
+    "react-dom": "^15.5.4"
+  }
+}
